requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pydantic[email]==1.10.7
python-dotenv==1.0.0
nltk==3.8.1
email-validator==2.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
aiofiles==23.1.0
httpx==0.24.1
markitdown[all]
scikit-learn==1.3.0
numpy<2.0.0
rank-bm25==0.2.2
redis==4.5.0
PyMuPDF==1.23.26
python-docx==1.1.0
weasyprint==60.0
fpdf2==2.7.6
jieba==0.42.1
